[MP-55] :recycle: 콘텐츠 수정 로직 변경

+++ src/main/java/com/pulse/content/application/service/ContentService.java
@@ -35,9 +35,12 @@ import com.pulse.content.domain.key.*;
 import com.pulse.content.domain.map.ContentHashTagMap;
 import com.pulse.content.domain.vo.ContentAttachment;
 import com.pulse.content.domain.vo.ContentDetail;
+import com.pulse.content.exception.ContentException;
+import com.pulse.content.exception.ErrorCode;
 import com.pulse.content.mapper.ContentMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -110,26 +113,24 @@ public class ContentService implements CreateContentsUseCase, FindContentUseCase
     @Override
     public UpdateContentResponseDTO update(UpdateContentRequestDTO updateContentRequestDTO) {
         ContentId contentId = updateContentRequestDTO.getContentId();
-        Content.contentIdValidation(contentId);
 
         // 콘텐츠 조회
         Content content = findContentPort.findContent(contentId);
-        Content.contentValidation(content);
+        if (ObjectUtils.isEmpty(content)) {
+            throw new ContentException(ErrorCode.ENTITY_NOT_FOUND);
+        }
 
         // 작성자와 수정자 아이디 비교
-        MemberId writerMemberId = content.getMemberId();
-        MemberId updateMemberId = updateContentRequestDTO.getMemberId();
-        Member.writerIdValidation(writerMemberId, updateMemberId);
+        MemberId updaterMemberId = updateContentRequestDTO.getMemberId();
+        content.writerIdValidate(updaterMemberId);
 
         // 콘텐츠 제목 및 내용 변경
         String title = updateContentRequestDTO.getTitle();
         String contentText = updateContentRequestDTO.getText();
-        Content.titleAndTextValidation(title, contentText);
-        content.putContentDetail(ContentDetail.of(title, contentText));
+        content.putContentDetail(title, contentText);
 
         // 콘텐츠 공개 범위 변경
         ContentVisibility contentVisibility = updateContentRequestDTO.getContentVisibility();
-        ContentVisibility.contentVisibilityValidation(contentVisibility);
         content.putContentVisibility(contentVisibility);
 
         // 콘텐츠 수정
@@ -140,10 +141,11 @@ public class ContentService implements CreateContentsUseCase, FindContentUseCase
         // 해시태그 및 해시태그 맵
         List<String> hashTagNames = updateContentRequestDTO.getHashTagNames();
         List<ContentHashTagMap> contentHashTagMaps = content.getContentHashTagMaps();
-        ContentHashTagMap.contentHashTagMapValidation(contentHashTagMaps);
 
-        // 해시태그 맵 삭제
-        deleteAllContentHashTagMap(contentHashTagMaps, hashTagNames);
+        if (!ObjectUtils.isEmpty(contentHashTagMaps)) {
+            // 수정한 해시태그 목록에 없는 해시태그맵 삭제
+            deleteAllContentHashTagMap(contentHashTagMaps, hashTagNames);
+        }
 
         // 수정한 해시태그 목록 중 신규 해시태그 등록
         createContentHashTagMap(contentHashTagMaps, hashTagNames, updatedContent);
